requirements.txt
paho-mqtt
pulsectl
